Removed debug output and changed packet output
This code is not in its finished state.  This code still requires that commands be sent to the RSA to retrieve data.  Also, this code is only designed to output the first three packets of the total BLE transmission for debug purposes.  This code will need to be edited to conform to the normal operations of the handheld translator