Initial Design for Handheld BLE Reciever
This is a stripped down version of the code we were supplied with.  I removed all of the serial outputs so that only the data itself is sent.  I changed the serial baud rate to match that of the Handheld.  I also removed the code that is used to interact with the BLE driver normally.  It is currently set to request a new BLE message every 5 seconds as a test.